+++ hypokrates/vocab/constants.py
@@ -9,6 +9,8 @@ from hypokrates.pubmed.constants import (
 )
 
 RXNORM_DRUGS_ENDPOINT = "/drugs.json"
+RXNORM_RXCUI_ENDPOINT = "/rxcui.json"
+RXNORM_ALLRELATED_ENDPOINT = "/rxcui/{rxcui}/allrelated.json"
 
 MESH_DATABASE = "mesh"
 
@@ -17,6 +19,8 @@ __all__ = [
     "ESEARCH_ENDPOINT",
     "ESUMMARY_ENDPOINT",
     "MESH_DATABASE",
+    "RXNORM_ALLRELATED_ENDPOINT",
     "RXNORM_DRUGS_ENDPOINT",
+    "RXNORM_RXCUI_ENDPOINT",
     "TOOL_NAME",
 ]